refactor(rework): extract textually-identical math helpers into reworkMathCore
